perf(test): speed up PieceDetail suite ~37% via WorkflowState stub
- Mock WorkflowState in PieceDetail.test.tsx with a lightweight stub
  (no autosave timer, no GlobalEntryField async calls). Tests that need
  real WorkflowState internals belong in WorkflowState.test.tsx.
- Move the "stale props" stale-props test to WorkflowState.test.tsx where
  it tests real WorkflowState behavior, not the stub.
- Replace userEvent.click → fireEvent.click/mouseDown for all simple
  button clicks (Share, Copy link, Stay, Leave, Browse location,
  gallery open). Add waitFor after the MUI Autocomplete Open button.
- Drop unused userEvent import.

Suite wall time: ~14s → ~10s (44 tests, up from 45 after move).
